+++ src/layers/6_client/fluent.ts
@@ -1,6 +1,6 @@
 import type { Anyware } from '../../lib/anyware/__.js'
 import type { Fluent } from '../../lib/fluent/__.js'
-import type { RequestCore } from '../../requestPipeline/__.js'
+import type { RequestPipeline } from '../../requestPipeline/__.js'
 import type { Schema } from '../../types/Schema/__.js'
 import type { Extension } from './extension/extension.js'
 import type { Config } from './Settings/Config.js'
@@ -50,7 +50,7 @@ export interface ClientContext {
   name: string
   input: InputStatic
   config: Config
-  retry: Anyware.Extension2<RequestCore.Core, { retrying: true }> | null
+  retry: Anyware.Extension2<RequestPipeline.Core, { retrying: true }> | null
   extensions: Extension[]
   scalars: Schema.Scalar.Registry
 }
