+++ src/layers/6_client/properties/anyware.ts
@@ -1,6 +1,6 @@
 import type { Anyware, Anyware as AnywareLib } from '../../../lib/anyware/__.js'
 import type { Fluent } from '../../../lib/fluent/__.js'
-import type { RequestCore } from '../../../requestPipeline/__.js'
+import type { RequestPipeline } from '../../../requestPipeline/__.js'
 import { createExtension } from '../extension/extension.js'
 import { defineProperties, type FnParametersProperty } from '../fluent.js'
 
@@ -14,13 +14,13 @@ export interface Anyware<$Args extends FnParametersProperty> {
    * TODO Anyware Docs.
    */
   (
-    anyware: AnywareLib.Extension2<RequestCore.Core<$Args['state']['context']['config']>>,
+    anyware: AnywareLib.Extension2<RequestPipeline.Core<$Args['state']['context']['config']>>,
   ): Fluent.IncrementNothing<$Args>
 }
 
 export const anywareProperties = defineProperties((builder, state) => {
   return {
-    anyware: (anyware: Anyware.Extension2<RequestCore.Core>) => {
+    anyware: (anyware: Anyware.Extension2<RequestPipeline.Core>) => {
       return builder({
         ...state,
         extensions: [...state.extensions, createExtension({ name: `InlineAnyware`, onRequest: anyware })],
