+++ src/layers/6_client/properties/retry.ts
@@ -1,6 +1,6 @@
 import type { Anyware } from '../../../lib/anyware/__.js'
 import type { Fluent } from '../../../lib/fluent/__.js'
-import type { RequestCore } from '../../../requestPipeline/__.js'
+import type { RequestPipeline } from '../../../requestPipeline/__.js'
 import { defineProperties, type FnParametersProperty } from '../fluent.js'
 
 export interface FnRetry extends Fluent.FnProperty<`retry`> {
@@ -12,12 +12,12 @@ export interface Retry<$Args extends FnParametersProperty> {
   /**
    * TODO Retry Docs.
    */
-  (extension: Anyware.Extension2<RequestCore.Core, { retrying: true }>): Fluent.IncrementNothing<$Args>
+  (extension: Anyware.Extension2<RequestPipeline.Core, { retrying: true }>): Fluent.IncrementNothing<$Args>
 }
 
 export const retryProperties = defineProperties((builder, state) => {
   return {
-    retry: (anyware: Anyware.Extension2<RequestCore.Core, { retrying: true }>) => {
+    retry: (anyware: Anyware.Extension2<RequestPipeline.Core, { retrying: true }>) => {
       return builder({ ...state, retry: anyware })
     },
   }
